Don't enforce and/or assume ReadUntil() responses have a LF suffix
The original telnetlib based approach using read_until() worked based on the
fact that the string returned by read_until() always had a trailing LF (\n).
This isn't the case with the way the response is returned by the raw socket
based implementation (ie. using string.partition()), so don't add a trailing
LF to the returned string anymore and fix up all occurances comparing with
a string containing a trailing LF.

(On a side note, it is simply wrong to statically append a LF where it would
rather be correct to append the separator, but well...)